Only check for clashing discriminator in same tree
This commit also fixes the discriminatorColumn check to check all the
entities it should be checking. Only the parent in an STI setup has the
"STI" inheritancePattern, the children just have the "entity-child"
tableType. Before two children with the same discriminatorValue would be
(incorrectly) allowed as they'd never even get checked.