docs: fix incorrect "chunk" description
The number to pass into "chunk" is *not* the number of chunks you desire, it is the desired size
of each chunk.

I lost a fair bit of time to this, because I only read the first line of the comment, and didn't
realize it conflicted with the example presented later in the comment.