Fix issue with arguments being inserted
According to https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/splice the splice method with a 0 delete will insert, where as a 1 will replace, this has been causing me massive headaches when i'm using 4-5 transaction repo's in a method, and have an argument that comes -after- the methods.  this leads to having a bunch of undefined arguments in the function calls.