fix: prevent dangling transactions
when you release the sql runner and had an open transactoin
we wouldn't actually do anything with it. in some scenarios,
releasing the runner closes the connection which causes
the transaction to roll back.

to keep the behavior uniform across drivers and scenarios,
release now explicitly rolls back transactions if they're open

fixes #3919