refactor: driver: Move [Delete/Insert/Update]Result processing to QueryRunners
While SELECT queries tend to result in objects containing columns and their values,
DELETE/INSERT/UPDATE queries all have different outputs depending on the driver. This
moves the processing of those outputs and their mapping to the [Delete/Insert/Update]Result
objects to the driver QueryRunners so that it can be altered without changing the QueryBuilders.