refactor: metadata: EntityManager.extractColumnsInEntity()
More efficient method of finding the columns that are contained in a given entity object.

Previously the object was recursively scanned for keys and all were converted to property paths,
then for every property path the columns were searched to find a matching one. Now the columns are
selected as the object is scanned, and since ownColumns is searched instead of all columns it is
more efficient.