refactor: driver: Add hasIndexedParameters for UPDATE parameters reordering
For certain drivers where the native parameters are not indexed (e.g. MySQL, "?")
the parameters which are found at the start of a query - specifically in UPDATE
SET param = ? - must be placed at the start of the native parameters objects. For
drivers where the parameters are indexed, the order must match the order of param
variable creation (e.g. Postgres, ":1").

This allows each driver to specify whether it cares about parameter order, allowing
UpdateQueryBuilder to reorder params if necessary.