Fix bug for Revert in Migration API Example
Migration will run successfully but will break when trying to revert. Here is the error
```
Error during migration revert:
Error: Supplied foreign key was not found in table question
```

The reason is because incorrect table is referenced when dropping the foreign key in the Migration's revert code (i.e `question` instead of `answer`).