fix: transformers now work on Update/CreateDateColumn
Because the transformer was being applied _after_ special handling logic
for dates, typeorm was comparing a date object to a string object and
detecting a change to the data.  This commit flips the logic so that
the transformer (if any) is applied prior to normalization, so that we
compare two date strings correctly when there is a transformation
supplied.

Fixes #7150

add explicit test case