Remove unused type variable from examples
The documentation on [embedded entities](https://typeorm.io/#/embedded-entities) for some reason names a `type` identifier in the arrow syntax. Linters (and some TypeScript compilers) will show error such as: "'type' is declared but its value is never read." (ts) or "'type' is defined but never used." (eslint). Furthermore, `type` is a TypeScript keyword, so best to avoid it as variable name.

Apparently, it is not needed per the last comment in #2915 , so simply replacing it with `()`