refactor: remove the need for a metadata table when using generated columns in postgres
The original implementation for generated columns required us to save the user provided "AS" expression into a metadata table. The reason behind it is that we are not able to get the original user provided expression from Postgres itself. This pr uses a different approach. Instead of retrieving the original expression, we just format and convert the user provided expression. With this, we are able to compare both expressions without the need for a metadata table