fix: add SQLite boolean escaping
When using `someRepository.findOneBy({ someField: true })`, `nativeParameters` has no keys and the query doesn't work properly, returning empty array. We updated to last TypeORM version (from 0.2.30) and this seems to solve it, but I don't really understand why booleans worked and now don't. If it means anything, we are using Cordova to develop and Android app. Anyways, this commit fixes it.