refactor: extract repeated predicate to isEntityLike method
This predicate occurs a couple of times in getEntityValue method.
I tried to deduce its value (isEntityLike) from its usage.

This refactor is not 100% pure, because I incorporated

    !(val instanceof Date)

expression into the method, even though it was mentioned in 2 out of 4
places that were replaced with the method call.

Again, studying the usage of the predicate, I decided that making sure
if the value is not a Date will only make the assumption more correct in
those 2 other places, because next operations assume that the value is
an entity.