docs: migration from .sql file
to polish off #1366 for good. noticing that users created a whole npm library for this problem to work around `multipleStatements` option

I wasn't sure whether to put this in the `QueryRunner` docs or here - it seems this should ideally only be used in migrations. Perhaps there are runtime applications for a psuedo-ETL type scenario, but it seems ill advised to even document the ability to do this at runtime?

There may be other edge cases where this fails, I've only tested it with standard pgdump style files with multi-statement, single statement-per-line