Package Upgrade: Upgrade the MongoDB driver to >6.x
The return value of `findOneAndX` family of MongoDB methods
now includes `null` when no document is found. This is a
breaking change from the previous behavior of returning
a result with metadata describing when a document is not found.
See: https://github.com/mongodb/node-mongodb-native/blob/HEAD/etc/notes/CHANGES_6.0.0.md