fix: Use BEGIN IMMEDIATE with pooled sqlite driver (#11)
### Description of change

Use `BEGIN IMMEDIATE` when opening transaction with pooled sqlite
driver. If we don't lock the DB immediately, it's possible to suffer
from DB locked errors in the middle of a transaction, which could
require retrying the entire transaction. It's better to handle these in
the beginning of a trx and do the retry there.