feat: Implement sqlite driver that uses single write connection and a pool of read connections (#12)
### Description of change

We have already implemented pooling for the SQLite driver. That uses a
single connection pool where connections are acquired from. Because
SQLite only allows a single writer to the DB at a time, other connection
will get DB locked error if they try to write at the same time. To
overcome this we have implemented retrying, which try again until the
operation succeeds (or times out). This is however suboptimal use of
resources.

A better model for this is to use a single write connection and a pool
of read connections and use the correct connection based on which is
needed. This is what this PR implements.

This basically decouples the lifecycle of a DB connection from the
lifecycle of a QueryRunner. QueryRunner's public interface basically has
2 ways which typeorm uses to interact with it
- Transaction methods: startTransaction, commitTransaction &
rollbackTransaction
- Query method: query

When queries are executed outside of a transaction, the query method is
called directly. In this case we 1) acquire either a read or write
connection 2) run the query and 3) release the connection back to the
pool. When a transaction is started, we 1) acquire a write connection 2)
run any queries using that connection 3) commit/rollback using that
connection and 4) release the write connection back to the pool.

The write connection uses a mutex from async-mutex to control access to
the write connection. This makes the resource model much more efficient,
as we don't need poll the DB if we are ready to write there again. This
seems to work very nicely.

With this change, I don't see any obstacles why a different SQLite
library wouldn't work as well. The polling model

**Executing a query outside of a transaction**


![image](https://github.com/n8n-io/typeorm/assets/10324676/bc42c42e-a3ac-4372-bbff-b281a1131214)


**Running a transaction**


![image](https://github.com/n8n-io/typeorm/assets/10324676/991f8499-dbcd-439b-aac4-c1a682037609)

### Performance

I ran some simple benchmarks with the n8n application. Workflow
execution using webhooks gives a slightly higher throughput (1070 reqs
vs 1105 reqs) (calling a webhook a WF that has few nodes). A combination
of writes and reads however gives much higher performance (5477 reqs vs
6710 reqs) (repeat 9 requests of WF execution with webook, 1 GET request
to get 10 executions of that WF).