Removed hardcoded database for generated column metadata
Setting a database on the inserts/deletes to typeorm_metadata can result in unnecessary migration regeneration for generated columns if the database name changes. This can occur if a dump of database `production`, for example, is restored to a database named `local`, and one attempts to generate migrations (typically for a different schema change) against `local`.

The "hardcoded" database name, `production`, will result in the generated column being dropped and recreated, potentially causing an incident if the change is not caught in development.