test: fix multi-database-basic-functionality
This test was not written properly, causing persistent changes to the process.platform which was breaking other tests. It was also not testing "darwin" properly.

Change the code to properly restore the process.platform when it's finished, as well as testing "darwin". Remove the filepath comparison when testing non-Windows behavior, where the filepath is used as-is.

Also update required rimraf usage for globs.

Closes: #11272