test: windows support for issue #8975 test
This test was setting the executable flag on the cli.js file and then running it. That doesn't work on Windows.

Improve cross-platform build support. Change the cliPath to use `node`. Remove the `chmod` code as that is not necessary if calling node directly.

Change fs.rmdir to fs.rm (other is deprecated).