fix(persistence): prevent save() from setting undefined fields to null on update
Fixes #11419. Previously, when using save() to update an entity, any nullable columns omitted (set as undefined) would be overwritten with null in the database. This change ensures that undefined fields are only set to null during insert, not update, so updates with undefined fields leave existing values unchanged. Adds/relies on tests in null-and-default-behaviour and issue-2331 to verify correct behavior. See: https://github.com/typeorm/typeorm/issues/11419

BREAKING CHANGE: Applications relying on save() setting undefined fields to null on update will need to explicitly set fields to null if that is desired.